requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.4
